hook up preset to analog channel editor; refine layout of sample rate and channel selector
